Specify slideNumberFormat option in options
This controls the slide number that is displayed in the lower right-hand
side of the slide. To make this clear to the user so that they know
how to modify this explicitly pass the default slide number format of

'%current% / %total%'

c.f. https://github.com/gnab/remark/wiki/Configuration